requirements.txt
pytest>=3.3.2
pytest-cov>=2.5.1
GitPython>=2.1.8
